# Improve performance of `unions`