This implements `unions` using `unionsWith` from the underlying
representation.

Before:

    λ mconcat $ take 100000000 $ cycle [singleton ()]
    fromOccurList [((),100000000)]
    (36.16 secs, 37,600,131,760 bytes)

After:

    λ mconcat $ take 100000000 $ cycle [singleton ()]
    fromOccurList [((),100000000)]
    (7.48 secs, 20,800,129,488 bytes)