@twanvl 2018-09-14
It is surprising that this makes such a huge difference in performance. If you look at how `unionsWith` is implemented in Data.Map, it is doing exactly the same thing. It probably comes down to inlining. Also: I don't trust benchmarks in ghci, since they are run without optimizations and might even involve some interpreted code.

Instead of {#- INLINE #-}, we should probably use {-# INLINABLE #-} pragmas on most functions.