@fredefox 2018-09-14
> I don't trust benchmarks in ghci, since they are run without optimizations

This is a good point. I think you're right that inlining is probably the main culprit. I figured that re-using the functionality from `Data.Map.Strict` couldn't hurt in any case. The maintainers probably know their stuff ;)

> Is this pragma necessary? I would think that ghc can figure this out itself.

I don't know how much GHC can figure out by itself. I kinda unintentionally appended this to the PR. (I'm doing benchmarking on some code that uses `multiset`.) My reason for adding the pragma was to mirror `Semigroup` -- I then found out that only `mappend` has an inline pragma in the class declaration and not `(<>)` ([Question on reddit](https://www.reddit.com/r/haskell/comments/9bveu0/monthly_hask_anything_september_2018/e5yj4by)). Another effect of the inline pragma is that cost centres are not created with `-fprof-auto`. But it may be contested if this is a somewhat selfish use-case.