@fredefox 2018-09-17
You were of course right that this was due to some of the code being run interpreted.

Previous:

```
	Mon Sep 17 11:47 2018 Time and Allocation Profiling Report  (Final)

	   multiset-bench +RTS -p -RTS

	total time  =        3.65 secs   (3655 ticks @ 1000 us, 1 processor)
	total alloc = 12,000,039,096 bytes  (excludes profiling overheads)

COST CENTRE    MODULE        SRC                                %time %alloc

union          Data.MultiSet Data/MultiSet.hs:403:1-52           46.5   53.3
heavy          Main          Main.hs:13:1-55                     23.1   46.7
foldlStrict    Data.MultiSet Data/MultiSet.hs:(714,1)-(717,62)   19.5    0.0
foldlStrict.z' Data.MultiSet Data/MultiSet.hs:717:21-30          10.9    0.0
```

With my changes:

```
	Mon Sep 17 11:43 2018 Time and Allocation Profiling Report  (Final)

	   multiset-bench +RTS -p -RTS

	total time  =        4.00 secs   (4002 ticks @ 1000 us, 1 processor)
	total alloc = 20,800,039,080 bytes  (excludes profiling overheads)

COST CENTRE MODULE        SRC                        %time %alloc

mconcat     Data.MultiSet Data/MultiSet.hs:196:5-20   83.8   73.1
heavy       Main          Main.hs:13:1-55             16.2   26.9
```

And for completeness the benchmark was:

```
module Main (main) where

import Control.Monad (void)
import Control.DeepSeq (NFData, force)
import Control.Exception (evaluate)

import Data.MultiSet

compute :: NFData a => a -> IO ()
compute = void . evaluate . force

heavy :: MultiSet ()
heavy = mconcat $ take 100000000 $ cycle [singleton ()]

main :: IO ()
main = compute heavy
```