@fredefox 2018-09-17
But doesn't it still seem weird to you that we need to allocate 12 GB to essentially compute `[((), 100000000)]`?